Cumulus-2470 Adds /s3credentials endpoint for cumulus distribution (#2261)
* Cumulus-2470 Adding /s3credentials endpoint to distribution API

* Cumulus-2470 exchanging shared code

* Cumulus-2470 lint fixes

* debugging

* debugging

* Fixing unit tests

* Cumulus-2470 Fixing unit tests

* Cumulus-2470 Add env variables to API

* Adding env variables

* Cumulus-2470 fixing env variables

* Cumulus-2470 Adding iam role permissions

* Cumulus-2470 Fixing lint errors

* Cumulus-2470 fixing unit tests

* Cumulus-2470 Fixing Unit tests

* Checking for CMR ACLs

* Fixing lint errors

* Addressing PR comments

* removing extra code

* Adding env variable

* Cumulus-2470 Addressing PR comments

* Cumulus-2470 Adding OAUTH_PROVIDER to tea

* Cumulus-2470

* Clean up
