+++ packages/api/app/distribution-routes.js
@@ -1,7 +1,6 @@
 const router = require('express-promise-router')();
 const { randomId } = require('@cumulus/common/test-utils');
 const { RecordDoesNotExist } = require('@cumulus/errors');
-
 const {
   handleLoginRequest,
   handleLogoutRequest,
@@ -10,39 +9,13 @@ const {
 } = require('../endpoints/distribution');
 const displayS3CredentialInstructions = require('../endpoints/s3credentials-readme');
 const { isAccessTokenExpired } = require('../lib/token');
-const { getConfigurations } = require('../lib/distribution');
+const { handleCredentialRequest } = require('../endpoints/s3credentials');
+const {
+  getConfigurations,
+  isPublicRequest,
+} = require('../lib/distribution');
 
 const version = require('../endpoints/version');
-
-/**
- * Helper function to pull bucket out of a path string.
- * Will ignore leading slash.
- * "/bucket/key" -> "bucket"
- * "bucket/key" -> "bucket"
- *
- * @param {string} path - express request path parameter
- * @returns {string} the first part of a path which is our bucket name
- */
-function bucketNameFromPath(path) {
-  return path.split('/').filter((d) => d).shift();
-}
-
-/**
- * Reads the input path and determines if this is a request for public data
- * or not.
- *
- * @param {string} path - req.path paramater
- * @returns {boolean} - whether this request goes to a public bucket
- */
-function isPublicRequest(path) {
-  try {
-    const publicBuckets = process.env.public_buckets.split(',');
-    const requestedBucket = bucketNameFromPath(path);
-    return publicBuckets.includes(requestedBucket);
-  } catch (error) {
-    return false;
-  }
-}
 
 /**
  * Ensure request is authorized through EarthdataLogin or redirect to become so.
@@ -99,16 +72,14 @@ const profile = (req, res) => res.send('Profile not available.');
 
 const pubkey = (req, res) => res.status(501).end();
 
-const s3Credentials = (req, res) => res.status(501).end();
-
 router.get('/', handleRootRequest);
 router.get('/locate', locate);
 router.get('/login', handleLoginRequest);
 router.get('/logout', handleLogoutRequest);
 router.get('/profile', profile);
 router.get('/pubkey', pubkey);
-router.get('/s3Credentials', s3Credentials);
-router.get('/s3CredentialsREADME', displayS3CredentialInstructions);
+router.get('/s3credentials', ensureAuthorizedOrRedirect, handleCredentialRequest);
+router.get('/s3credentialsREADME', displayS3CredentialInstructions);
 // Use router.use to leverage custom version middleware
 router.use('/version', version);
