+++ tf-modules/cumulus_distribution/api.tf
@@ -3,15 +3,20 @@ locals {
   api_uri                   = var.api_url == null ? "https://${local.api_id}.execute-api.${data.aws_region.current.name}.amazonaws.com/${var.api_gateway_stage}/" : var.api_url
   api_redirect_uri          = "${local.api_uri}login"
   api_env_variables = {
-      AccessTokensTable     = aws_dynamodb_table.access_tokens.id
-      DISTRIBUTION_ENDPOINT = local.api_uri
+      AccessTokensTable              = aws_dynamodb_table.access_tokens.id
+      CMR_ACL_BASED_CREDENTIALS      = var.cmr_acl_based_credentials
+      CMR_ENVIRONMENT                = var.cmr_environment
+      DISTRIBUTION_ENDPOINT          = local.api_uri
       DISTRIBUTION_REDIRECT_ENDPOINT = local.api_redirect_uri
-      OAUTH_CLIENT_ID       = var.oauth_client_id
+      OAUTH_CLIENT_ID                = var.oauth_client_id
       OAUTH_CLIENT_PASSWORD_SECRETE_NAME = length(var.oauth_client_password) == 0 ? null : aws_secretsmanager_secret.api_oauth_client_password.name
-      OAUTH_HOST_URL        = var.oauth_host_url
-      OAUTH_PROVIDER        = var.oauth_provider
-      public_buckets        = join(",", var.public_buckets)
-      stackName             = var.prefix
+      OAUTH_HOST_URL                 = var.oauth_host_url
+      OAUTH_PROVIDER                 = var.oauth_provider
+      stackName                      = var.prefix
+      STS_CREDENTIALS_LAMBDA         = var.sts_credentials_lambda_function_arn
+      STS_POLICY_HELPER_LAMBDA       = var.sts_policy_helper_lambda_function_arn
+      cmr_provider                   = var.cmr_provider
+      public_buckets                 = join(",", var.public_buckets)
   }
   lambda_security_group_ids = [aws_security_group.no_ingress_all_egress[0].id]
 }
